requirements.txt
line-bot-sdk
flask==1.0.2
google-cloud-firestore==0.29.0
firebase-admin==2.13.0
